chore: support anomaly center in project detail
